breaking everything down to 4 separate files to compare
