+++ services/ui-src/src/measures/2022/shared/CommonQuestions/OptionalMeasureStrat/subCatClassification.tsx
@@ -1,7 +1,7 @@
 import * as QMR from "components";
 import * as CUI from "@chakra-ui/react";
 import { useFieldArray, useFormContext } from "react-hook-form";
-import { NDRSets } from "./ndrSets";
+import { NDRSets } from "./NDR/ndrSets";
 
 interface AddAnotherButtonProps {
   /** onClick state updating function for dynamic rendering */
