+++ services/ui-src/src/measures/2023/shared/CommonQuestions/OptionalMeasureStrat/omsNodeBuilder.tsx
@@ -5,7 +5,7 @@ import { OmsNode } from "./data";
 
 import { AddAnotherSection } from "./additionalCategory";
 import { SubCatSection } from "./subCatClassification";
-import { NDRSets } from "./ndrSets";
+import { NDRSets } from "./NDR/ndrSets";
 
 interface CheckboxChildrenProps extends OmsNode {
   /** name for react-hook-form registration */
